scripts/requirements-test.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pyyaml>=6.0
coverage>=7.3.0
